Separate EntityExplosion in DRE
